backend/pyproject.toml
[project]
name = "orbis-backend"
version = "0.1.0"
description = "Orbis — CV as GraphRAG"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "neo4j>=5.20.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3.0",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "anthropic>=0.40.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "cryptography>=43.0.0",
    "python-multipart>=0.0.9",
    "mcp[cli]>=1.0.0",
    "litellm>=1.52.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "respx>=0.22.0",
    "ruff>=0.7.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that call external LLM APIs (deselect with '-m \"not integration\"')",
]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "C90", # mccabe complexity
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
]
ignore = ["E501", "B008", "B905", "ARG001"]

[tool.ruff.lint.mccabe]
max-complexity = 30

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
